requirements.txt
openpyxl>=3.1.2
python-pptx>=0.6.23
pandas>=2.0.0
